API

Gatekeeper Formation Flight

Public:

GatekeeperFormationFlight.Cardioid — Type

Cardioid(x, y, Rmin=0.0, Rmax=0.15)

Construct a Cardioid WEZ at position x, y with radii Rmin, and Rmax.

GatekeeperFormationFlight.Cbez — Type

Cbez(x, y, ψ)
Cbez(x, y, ψ, μ=0.9, ā=0.25, v=1.0, R=π/2, t=π/2)

Define a Curve-Only Basic WEZ at some x, y with a heading ψ.

GatekeeperFormationFlight.CircularWez — Type

CircularWez(x, y, R=0.15)

Define a Circluar WEZ at some x, y with radius R.

GatekeeperFormationFlight.DubinsRRTProblem — Type

DubinsRRTProblem(domain, turning_radius, wezes)

Construct a dubins RRT* problem.

Parameters:

domain: a tuple defining the domain

turning_radius: turning radius for Dubins paths

wezes: A vector of wezes to avoid

Example:

using GatekeeperFormationFlight
using StaticArrays

min_domain = SVector{3}(0.0, 0.0, -1.0*π)
max_domain = SVector{3}(1.0, 1.0, 1.0*π)
domain = (min_domain, max_domain)

turning_radius = 0.1

wezes = [Cardioid(rand(), rand()) for i=1:10]

rrt_problem = DubinsRRTProblem(domain, turning_radius, wezes)

GatekeeperFormationFlight.GatekeeperProblem — Type

GatekeeperProblem(kwargs...)

construct a GatekeeperProblem. The arguments are:

wezes, # list of wezes

reference_path, # path of the leader

offset = SVector(0.0, 0.0, 0.0), # desired offset from the leaders path

turning_radius = 0.1, # max turning radius of a robot

maxTshorizon = 1.0, # maximum switching time

reconnectionstepsize = 0.1, # resolution used for checking reconnection point

collisioncheckstep_size = 0.01, # resolution used for checking collision along a path

integrationmaxstep_size = 0.05, # max integration step size in nominal tracking

switchstepsize = 0.05 # resolution used to decrease switch time

GatekeeperFormationFlight.Robot — Type

Robot(x, y, ψ)
Robot([x, y, ψ])

Create a robot with some x, y position and heading ψ.

GatekeeperFormationFlight.apply_input_bounds — Method

apply_input_bounds(v, ω; v_min = 0.8, v_max = 1.0, R_min = 0.1)

apply input bounds to a v, ω. Extracted as a function to make sure input bounds get applied everywhere uniformly.

GatekeeperFormationFlight.collision_distance — Function

collision_distance(wez::AbstractWez, robot::Robot)
collision_distance(wezez::Vector{AbstractWez}, robot::Robot)

returns the distance to collision for this robot to the wez. Positive if safe. If the robot is within the wez, it will return a negative number.

GatekeeperFormationFlight.construct_candidate_trajectory — Method

construct_candidate_trajectory(time, state, prob::GatekeeperProblem)

constructs a nominal trajectory, and working backwards tries to construct backup trajectories. If a safe trajectory is found, this is returned as the candidate trajectory. if not, nothing is returned.

GatekeeperFormationFlight.get_reference_state — Method

get_reference_state(path::Vector{DubinsPath}, time, offset)

get the x_d to follow a path at some time, with a given offset.

GatekeeperFormationFlight.get_reference_state — Method

get_reference_state(path::Vector{DubinsPath}, time)

get the x_d to follow a path at some time.

GatekeeperFormationFlight.get_reference_state_and_input — Method

get_reference_state_and_input(path::Vector{DubinsPath}, time, offset)

get the (x_d, u_d) to follow a path at some time, with a given offset.

GatekeeperFormationFlight.get_reference_state_and_input — Method

get_reference_state_and_input(path::Vector{DubinsPath}, time)

get the (x_d, u_d) to follow a path at some time

GatekeeperFormationFlight.is_colliding — Function

is_colliding(reg1::CollisionRegion, reg2::CollisionRegion, tol=0)

return true if the distance between reg1 and reg2 is less than tol.

GatekeeperFormationFlight.is_colliding — Function

is_colliding(wez::AbstractWez, robot::Robot, tol)
is_colliding(wezes::Vector{AbstractWez}, robot::Robot, tol)

returns collision_distance(wez, robot) <= tol

GatekeeperFormationFlight.is_colliding — Method

is_colliding(wezes::Vector{AbstractWez}, path::DubinsPath, tol=1e-5)

returns true if the dubins path gets within tol distance of any wez. See docs for is_colliding(wez, path, tol). Sorts the wezes by distance before checking each one.

GatekeeperFormationFlight.is_colliding — Method

is_colliding(wez::AbstractWez, path::DubinsPath, tol=1e-5)

returns true if the dubins path gets within tol distance of the wez.

Checks by sampling every point tol apart along the path.

To reduce computational cost, it checks bounding boxes first, and then also skips points too close to each other where there is no possibility of collision.

GatekeeperFormationFlight.minimum_distance — Function

minimum_distance(c::AbstractWez p::SVector{2})

utility function to get the smallest distance to a wez from the current location, indepednent of robot orientations. Currently only implemented for Cardioid and CircularWez.

GatekeeperFormationFlight.minimum_distance — Method

minimum_distance(reg1::CollisionRegion, reg2::CollisionRegion)

return the minimum distance between reg1 and reg2

GatekeeperFormationFlight.simulate_closed_loop_gatekeeper — Method

simulate_closed_loop_gatekeeper(initial_state, tspan, prob::GatekeeperProblem)

returns the closed-loop trajectory from running gatekeeper

GatekeeperFormationFlight.total_path_length — Method

total_path_length(path::DubinsPath)

returns the total path length of a dubins path

GatekeeperFormationFlight.total_path_length — Method

total_path_length(path::Vector{DubinsPath})

returns the total path length of a vector of DubinPath

GatekeeperFormationFlight.tracking_controller — Method

tracking_controller(x, x_d, u_d; k1=450.0, k2=200.0, k3=50.0)

returns the [v, ω] to track a state x_d from current state x, given a feedforward term u_d.

GatekeeperFormationFlight.wez_coordinates — Function

wez_coordinates(θ, wez::AbstractWez, robot::Robot)

returns the (x, y) coordinates of the wez boundary relative to a robot, parameterized by the polar angle θ ∈ [0, 2π].

Private:

GatekeeperFormationFlight.Circle — Type

Circle(center, radius)

defines a circluar region by center and radius.

GatekeeperFormationFlight.CompositeTrajectory — Type

CompositeTrajectory(nominal_trajectory, backup_trajectory, switch_time)

defines a composite trajectory that includes a nominal trjectory, a backup trajectory, and a switch time

GatekeeperFormationFlight.LineSegment — Type

LineSegment(p1, p2)

defines a line segment connecting points p1 and p2. Both p1 and p2 must be of type SVector{2, F}.

StaticArraysCore.SVector — Method

SVector(r::Robot)

returns a SVector{3, F} containing [x, y, ψ]

GatekeeperFormationFlight.RRTStar.collision_free — Method

collision_free(problem, x_nearest, x_new; step_size=0.001)

checks if a path from x_nearest to x_new is collision free with respect to the wezes in problem.wezes.

GatekeeperFormationFlight.RRTStar.near — Method

    near(problem, nodes, x_new)

returns the set of indices of nearby nodes to a state x_new. For simplicity, it currently just returns all nodes.

GatekeeperFormationFlight.RRTStar.nearest — Method

nearest(problem, nodes, x_rand)

returns the index of the nearest node within the list of nodes to the state x_rand, according to a dubins path distance.

GatekeeperFormationFlight.RRTStar.path_cost — Method

path_cost(problem, x_near, x_new)

returns the dubins path length from xnear to xnew.

GatekeeperFormationFlight.RRTStar.sample_free — Method

sample_free(problem::DubinsRRTProblem)

returns a random state within the problem domain

GatekeeperFormationFlight.RRTStar.steer — Method

steer(problem, x_nearest, x_rand; max_travel_dist=0.2)

returns the state by travelling upto max_travel_dist towards x_rand starting from x_nearest.

GatekeeperFormationFlight.closed_loop_tracking_backup! — Method

closed_loop_tracking_backup!(D, state, backup_path::Vector{DubinsPath}, backup_time)

defines the closed_loop dynamics for tracking the backup trajectory

GatekeeperFormationFlight.closed_loop_tracking_composite! — Method

closed_loop_tracking_composite!(D, state, params, time)

defines the closed_loop dynamics for tracking a composite trajectory

GatekeeperFormationFlight.closed_loop_tracking_nominal! — Method

closed_loop_tracking_nominal!(D, state, prob::GatekeeperProblem, time)

defines the closed_loop dynamics for tracking the nominal trajectory

GatekeeperFormationFlight.construct_candidate_backup_trajectory — Method

construct_candidate_backup_trajectory(initial_state, prob::GatekeeperProblem)

from some initial state, this function construsts the best backup trajectory

GatekeeperFormationFlight.construct_candidate_nominal_trajectory — Method

construct_candidate_nominal_trajectory(time, state, prob::GatekeeperProblem)

does a closed-loop simulation of the nominal controller until either the time runs out or until it hits a wez

GatekeeperFormationFlight.construct_reconnection_paths — Method

construct_reconnection_paths(reconnection_sites, initial_state, turning_radius)

given a list of reconnection sites, and the state from which we want to start, this function returns a list of dubins paths that connect to the reconnection sites.

GatekeeperFormationFlight.construct_reconnection_sites — Function

construct_reconnection_sites(reference_path::Vector{DubinsPath}, reconnection_step_size=0.01)

along a reference path, define the set of points where a trajectory can possibly reconnect. Returns a Vector{Tuple(Int64, SVector{3, Float64})}. Each entry of this vector contains first the index of the reference path where the reconnection can happen, and second the state of reconnection.

GatekeeperFormationFlight.create_collision_region — Method

create_collision_region(c <: AbstractWez)

given a wez, create a maximum collision region around it. Currently only implemented for Cardioid.

GatekeeperFormationFlight.create_collision_region — Method

create_collision_region(path::DubinsPath, segment_id)

creates a collision region for the segment_id section of a dubins path, assuming it has a known segment type

GatekeeperFormationFlight.dubins_distance — Function

dubins_distance(q1, q2, turning_radius=0.1)

get the distance between two states, if connected by a dubins path

GatekeeperFormationFlight.get_ω_sign — Method

get_ω_sign(path::DubinsPath, index)

get the sign of the ω for the index-th subpart of a DubinsPath

GatekeeperFormationFlight.post_extrapolate — Method

post_extrapolate(path, time)

return the state, input for a reference path beyond the end of the path, assumes the robot moves with v=1.0, ω=0.0.

GatekeeperFormationFlight.pre_extrapolate — Method

pre_extrapolate(path, time)

return the state, input for a reference path for negative time. Assumes the robot moves with v=1.0, ω=0.0.

GatekeeperFormationFlight.terminate_condition — Method

terminate_condition(state,time,integrator)

callback to terminate integration when there is a collision

GatekeeperFormationFlight.update_committed_affect! — Method

update_committed_affect!(integrator)

callback to recompute the committed trajectory. Will replace the integrator's parameter if a new committed trajectory is found

GatekeeperFormationFlight.wrapToPi — Method

wrapToPi(θ)

returns θ wrapped to lie within (-π, π]

RRTStar

Public:

GatekeeperFormationFlight.RRTStar.Node — Type

Node{T}(state::T, parent_index::Int64, incremental_cost::Float64)

Defines a node in the RRT* Tree, with a defined state, parent_index, and the incremental_cost from the parent to the current node. Since the nodes are stored in a vector, the root node should have a parent_index=0.

GatekeeperFormationFlight.RRTStar.get_best_path — Method

get_best_path(problem::AbstractProblem, nodes::Vector{Node{T}}, x_goal::T)

returns the (true, best_path) if a path to x_goal is found. returns (false, nothing) if no path to x_goal is found.

The best path is a list of nodes from the root to a node on the tree. returns the states along the nodes.

GatekeeperFormationFlight.RRTStar.rrt_star — Method

rrt_star(problem::AbstractProblem, nodes::Vector{Node}, max_iters::Int64; do_rewire=true)

Adds max_iters nodes to the RRT* star. Will modify the nodes vector.

Private:

GatekeeperFormationFlight.RRTStar.cost — Method

cost(i::Int64, nodes::Vector{Node})

returns the cost to get to node i. Computes this quantity recursively.